move powspace, linespace to utils
