+++ torchinductor/triton_ops/bench/bench_conv.py
@@ -1,29 +1,14 @@
-import math
 import torch
 import triton
 
 import torchinductor.triton_ops
-
-
-def rounded_linspace(low, high, steps, div):
-    ret = torch.linspace(low, high, steps)
-    ret = (ret.int() + div - 1) // div * div
-    ret = torch.unique(ret)
-    return list(map(int, ret))
-
-
-def powspace(start, stop, pow, steps):
-    start = math.log(start, pow)
-    stop = math.log(stop, pow)
-    ret = torch.pow(pow, torch.linspace(start, stop, steps))
-    return list(map(int, ret))
-
+import utils
 
 # conv benchmarks
 conv_confs = [
     triton.testing.Benchmark(
         x_names=["IN_H", "IN_W"],
-        x_vals=[8, 16, 32, 64, 128, 256], #powspace(8, 256, 2, 1),
+        x_vals=utils.powspace(8, 256, 2, 1),
         line_arg="provider",
         line_vals=["cublas", "triton"],
         line_names=["cuBLAS", "Triton"],
@@ -32,8 +17,8 @@ conv_confs = [
         args={"BATCH": BATCH, "IN_C": IN_C, "KERNEL_N": KERNEL_N,
               "KERNEL_H": KERNEL_H, "KERNEL_W": KERNEL_W},
     ) for BATCH in [32]
-    for IN_C in [128]  #powspace(16, 256, 2, 1)
-    for KERNEL_N in [32]  #powspace(16, 256, 2, 1)
+    for IN_C in [128]  # powspace(16, 256, 2, 1)
+    for KERNEL_N in [32]  # powspace(16, 256, 2, 1)
     for KERNEL_H in [2]
     for KERNEL_W in [2]
 ]
@@ -76,4 +61,4 @@ def bench_op(
         return tflops(ms), tflops(max_ms), tflops(min_ms)
 
 
-bench_op.run(show_plots=True, print_data=True)
+bench_op.run(print_data=True)
